# src/main/java/com/openbanking/accounts/model/OBReadAccount6.java
package com.openbanking.accounts.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadAccount6
 */
@Getter
@Setter
public class OBReadAccount6 implements Serializable {
	private static final long serialVersionUID = 1L;

}